style: rename APIAuth -> IDAPIAuth
