+++ src/server/lib/idapi/user.ts
@@ -5,7 +5,7 @@ import {
   APIAddClientAccessToken,
   APIForwardSessionIdentifier,
   IDAPIError,
-} from '@/server/lib/APIFetch';
+} from '@/server/lib/IDAPIFetch';
 import { logger } from '@/server/lib/logger';
 import {
   ConsentsErrors,
