+++ src/server/routes/consents.ts
@@ -36,7 +36,7 @@ import {
 import { CONSENTS_PAGES } from '@/client/models/ConsentsPages';
 import { fourZeroFourRender } from '@/server/lib/middleware/404';
 import { handleAsyncErrors } from '@/server/lib/expressWrappers';
-import { IDAPIError } from '@/server/lib/APIFetch';
+import { IDAPIError } from '@/server/lib/IDAPIFetch';
 import { getConfiguration } from '@/server/lib/getConfiguration';
 import { Configuration } from '@/server/models/Configuration';
 import { PageTitle } from '@/shared/model/PageTitle';
